Implement triple-buffering + async. rendering
Enable triple-buffering, and swap buffers in the VSYNC callback.

This ensures that all frames will look perfect without tearing, without
having the problem of double buffering, namely the frame rate locked
to an integer division of 60, which is problematic for games that are
not fullspeed, and terrible in case of PAL games.

For NTSC games, the frame limiter is also now implemented as a wait for
the VSYNC interrupt.

Besides, the frames are now rendered asynchronously, to shave off a bit
of overhead.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>