Normalize line endings
Git repositories use the LF (UNIX) format for text files internally,
and will automatically convert them to CRLF (DOS) format when doing a
checkout or checkin of text files on Windows.

By default, git does not auto-detect text files (vs. binary files) so
this feature is disabled.

Force git to auto-detect text files, and normalize line endings to the
format it expects, so that all the files use the expected text format.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>